Consider skip in Hash_Block, no uneven skip/blocks
In R3-Alpha, the hash built to accelerate the set operations was not
accounting for the skip size.  Every item was hashed, meaning that
"values" could overwrite "keys" in the hash slots...despite the value
not corresponding to a position that could start a "full record":

https://github.com/rebol/rebol/blob/25033f897b2bd466068d7663563cd3ff64740b94/src/core/s-crc.c#L311

This accounts for that by taking the skip into account.  But there are
still a number of open questions about the meaning of /skip in these
operations.  To try and limit the problems at least somewhat, this also
adds an error disallowing the use of /skip sizes that the data is not
an even multiple of...as handling the arbitrary END markers in the
middle of data creates more complex code, and is usually probably a
bug in the caller anyway.