Eliminate Get_Var_Into parallel fetch code
With SELF not being an actual word that could be looked up to a
variable, it was not possible to take a word that was bound to index 0
and a context and hand back a pointer to an OBJECT!.  This meant
that if a variable lookup was going to consider SELF lookups it had to
use a special accessor called GET_VAR_INTO instead of just GET_VAR,
which took the word to look up and provided a destination to write to.

Given the performance-critical nature of GET_VAR, the idea of adding
more parameters and conditions to the function was not palatable, so
the two routines were essentially copies of each other with one small
difference.  However, now that SELF is an ordinary lookup and it's
always possible to get a pointer to a variable, only one routine is
necessary.

This also affords an enhancement to the infix lookahead, which is able
to get a pointer to the pointed-to next value and *then* decide if it is
going to extract a function from it or copy the value into the `out` cell
to use as a fetched word lookup.