Get rid of `;R3` comments in bootstrap scripts
There were several instances of `;R3` at the end of lines in the
make-boot files.  Presumably these are holdovers from a time when
Rebol2 was being used to build Rebol3...or transitioning away from
that.

Since the beginning of Ren-C, R3-Alpha builds are a minimal
requirement for bootstrap, so the comments are superfluous.