Sync LOCK, PROTECT meanings for series/typesets
Early in the change from keys being "unwords" to being typesets,
the change was significant...and the choice to not combine that with
renaming the flags EXT_WORD_HIDE and EXT_WORD_LOCK to
EXT_TYPESET_HIDE and EXT_TYPESET_LOCK was held off.

Additionally there was a question about how the words "LOCK" and
"PROTECT" were being used in series, which was different...
series used the word "PROTECT" to mean unmodifiable (in the
sense that EXT_WORD_LOCK was using) and then used LOCK to
mean "fixed size".  It wasn't clear if flags should adopt the PROTECT
or LOCK convention for unmodifiable.

Having been a while with the incorrect _WORD in the flag for a
typeset, it was time to make a decision.  The decision is to favor the
word LOCK for meaning Read-Only, and move the series concept of
being fixed size to SER_FIXED_SIZE.

The reason is because firstly, SER_LOCK doesn't communicate that
much intuitively to a reader who has to go look up exactly what that
means.  It seems less likely that it would mean "cannot change size"
than "cannot be modified", and SER_FIXED_SIZE is unambiguous,
so it seems preferable.

The reason for SER_LOCKED and EXT_TYPESET_LOCKED instead
of SER_PROTECT/ED and EXT_TYEPSET_PROTECT/ED is because
`protect` as a function represents multiple kinds of protection.  In fact,
hiding is done with `protect/hide`.  So being more specific than just
"protect" with the kind of protection makes sense.

<lock> is also a good short word that might be used as a tag on args
in functions... a kind of literate <const>.  So <lock> and <hide> are a
good fit for their shortness.

The longer and more literate forms of LOCKED and HIDDEN were
chosen to help communicate these are status bits, which makes
absorbing the code a little more easier at the cost of 2 characters.