LOGICAL() REBOOL coercion, trig cleanup
This is a first test of a new LOGICAL() macro, designed to be used in
the place of (REBOOL)style_casts.

It is used in the implementation of NOT(REF()), an alternative to
!REF() which is slightly more visible (like NOT_END() is somewhat
more scannable than !IS_END()).  But also, it allows for tests of
strange choices for the boolean type...such as having TRUE and
FALSE be pointer constants, to make type errors more obvious.

Whether using strange definitions for booleans will be a long term
thing or just a way of sweeping for bugs now and then will be seen,
but this is just adapting the trig functions to the new parameter format
for natives.