POKE on BITSET! only pokes LOGIC! values
R3-Alpha would allow you to poke numbers into bitsets, and checked
to see if those numbers were zero or not (where zero is false).  This
behavior of zero-detection as logic is used no other place in the code,
confuses the type handling, and isn't consistent with Rebol's "zero is
not false" belief.

    >> foo: make bitset! 1
    == make bitset! #{00}

    >> poke foo 0 0.000000000000001
    == make bitset! #{80}

    >> bar: make bitset! 1
    == make bitset! #{00}

    >> poke foo 0 0.000000000000000
    == make bitset! #{00}

It makes more sense to constrain to logic true and false than it does
to open the field to anything conditionally true by default.  (It's easy
enough to say TRUE? xxx or NOT ZERO? xxx to get a LOGIC!.)

So this reigns in the feature and makes anything poked into a bitset
that isn't a LOGIC! an error.