Make RXA_LOGIC support LOGIC!-in-header bit
The adjustments to the workings of LOGIC! to use the header bit
for conditional truth and falsehood were not accounted for by the
RXIARG interface.  Realizing that this actually proxied into a new
structure (vs. shadowing the bits as is done with REBSER) it was
trivial to add support for logic to the RXIARG in particular.

This was found during a codebase-wide audit of the use of REBOOL
and REBFLG.  Since there was only one occurrence in the core, this
tightens it so that it can pass the mechanism of that audit (which
differentiates between integers and booleans in terms of types).  It
has a somewhat aggressive assert that the integer value placed in
the slot is a 1 or a 0 as part of the protocol, which is not strictly
necessary but may be a way of catching confused code that thinks
it's putting a character in the payload (or something like that).