Boolean review changes to uni/string flags
In doing a review of the places where integers were being passed
to booleans or vice-versa, one area that had some confusing issues
was unicode encoding/decoding and related string routines.  The
function To_REBOL_Path in particular had an integer parameter
indicating unicode input that was sometimes passed as -1 to still
pass as "truthy" on being unicode input, but the negativeness as
a signal that even if the output only used Latin1 characters that the
result should still be created as a REBUNI wide string.  :-/

Given that was effectively a third option on the function, the options
were converted to bitflags.  Decode_UTF8 had a similarly strange
avoidance of an additional boolean in order to wedge a "2" into an
unrelated variable.

Other changes in the group of loosely related routines involve similar
clarifications, as well as renamings of flags like "uni" to "unicode",
given the common convention of naming REBUNI variables that hold
a character simply "uni".  (is_uni was considered, but unicode had
been used before and seemed to read better).  Variables with names
like "lee" or "le" or "ccr" were converted to "little_endian" and "crlf_to_lf"
or "lf_to_crlf" as appropriate.  They were not invoked that often and
required comments to explain them, when a self-explanatory name
would do better.

The MOPT_ANSI_ONLY flag had a "//parened" comment on it, which
led it to seem like it should be changed to MOPT_NON_ANSI_PARENED
instead.

Beyond that a few changes of REBFLG to REBOOL, in line with the
upcoming large-scale audit.