Fix precedence bug in file open permissions
The line:

    modes = O_BINARY | GET_FLAG(file->modes, RFM_READ) ? O_RDONLY : O_RDWR;

...has a precedence error, in that it is not adding either O_RDONLY
or O_RDWR to O_BINARY.  Rather it is OR'ing together O_BINARY
and the state of the RFM_READ flag...and then deciding to either
use O_RDONLY or O_RDWR, minus inclusion of O_BINARY.

Whether this winds up a problem or not depends on whether O_BINARY
is nonzero.  If it's 0, then it works correctly, if not it won't work.  It seems
to be 0 on linux, but Google searching reveals for instance on "sanos"
it is 0x8000:

http://www.jbox.dk/sanos/source/include/fcntl.h.html

This corrects the precedence using parentheses.