Remove unused Partial() argument
The fourth parameter to Partial was a boolean being passed in
as an integer 0.  But when looking to change those callsites to FALSE
instead, it turned out there were no cases of passing TRUE, and the
(somewhat confusing) parameter had no usages.

This removes the parameter and tidies Partial() slightly.