Tidy new-line implementation, no bitwise XOR
Bitwise XOR drew scrutiny on this routine which was XOR'ing a
boolean against an integral comparison result.  It's technically
legal, so long as the representation of TRUE used is exactly 1 and
the value of the REBOOL has been normalized to that.  But it's an
easy enough dependency to get rid of, and the code needed to be
updated to the new parameter scheme anyway.  So this tidies it to
be more obvious.