REBOOL fix to Temp_Byte_Chars_May_Fail
The Temp_Byte_Chars_May_Fail routine is a renaming of R3-Alpha's
somewhat confusing "Qualify_String".  It took a fourth parameter that
was a REBINT for "opts", yet all the callsites passed either FALSE or
TRUE.  The integer was checked to see if it was < 2 (it always was)
and then if it was 0 it would not accept UTF8 bytes > 0x80.

This makes the parameter a formal REBOOL and renames it simply
to `allow_utf8`.