Handle END or past-end cases for ?? in PARSE
The ?? debugging command in PARSE shows the current rule and
the current position in the data.  Both of these could be at the end of
series, which are not currently legal to pass to the molding of values.
This adds very basic handling for the end conditions in custom print
messages for those cases.