Fetch label for system functions in backtrace
When a system function like `make-port*` is used to do a dispatch,
it was previously the case that it treated it just like any anonymous
function value called from the C code.  However, since the function
var was obtained by frame index number, a frame key exists at that
index too.  This gets that key and uses its symbol to name the call
in the backtrace, which makes debugging a little easier.