Always push stacks in Do_Core(), filter backtrace
The idea that a call frame is the local variable state of Do_Core
enables another concept, which is that every call to Do_Core pushes
itself to the "Rebol call stack", whether an evaluation is in progress or
not...and stays on the stack until the end of the call. Then the "mode"
variable indicates what that level in the stack is doing...which may be
gathering state for a pending function call or just looping through and
doing word lookups or no-op evaluations.

As a result, this puts things like paren evaluations into the stack. The
benefit of this is that by being on the stack, the garbage collector can
know the evaluator is processing that array and protect it. This is
important in circumstances like:

    eval copy quote (recycle "this could be invalid...")

The alternative cost of protecting a value cell for something like `eval`
in this case creates overhead on every evaluation loop. Chaining the
stack level in and taking it out is something nearly every block or paren
had to do already (if it contained at least one function call) and may
have to do many times (if it had more than one). So paying for it once
each time and getting the fringe benefit of protecting temporary arrays
is a net win.

This is a simple version of the approach, which could go even farther by
actually allocating Reb_Call frames sequentially in a series (as opposed
to just a series of pointers), eliminating the ->prior pointer and using
plain pointer subtraction to navigate the fixed size array.  This would allow
stack depths more limited by heap memory for user code than by the size
of the compiled in C stack.  Performance impact should be studied
related to this to see if it's a good idea.