TITLE reflect of function doesn't VAL_TYPE(END)
The new rules for END markers are that there is no END! Rebol
type.  As such, VAL_TYPE of an END is not legal...if there is a risk
that a REBVAL could be an END that must be tested separately.

This corrects an instance of the rule being broken with the fetch of
the "TITLE" of a function in the reflection API, which also raises the
question of whether that should be a core service or a "user-mode"
and mezzanine protocol built on SPEC reflection only.