Stack balance Do_Construct w/unpaired SET-WORD!
Do_Construct() would push SET-WORD!s that it was planning
on setting once it found an inert value to assign to it.  However,
if SET-WORD!s trailed and reached the end of the block without
any value coming after it, the variables were left as their default
(currently NONE!)...and nothing cleared the pending sets from
the stack.  This caused the evaluator to complain that the native
did not return the data stack to its balanced position.

This resolves the issue with a DS_DROP_TO and lightly
simplifies and comments Do_Construct().