Optimize Find_Str_Char, fix negative skip
The Find_Str_Char() routine was taking up about 10% of the build time
for hostilefork.com using the Draem static website builder, and it
seemed a bit excessive considering what the routine did.  By making
the routine a little more sophisticated and leaning on C library
routines for the common case of finding 1 or 2 characters in a byte
string (that aren't NUL)--when searching forward and skipping by 1
character at a time--this makes it faster.

For the moment the old routine is called as a double-check in debug
builds just to make sure nothing has changed.  However, this called
out a problem with infinite loops based on the sign of the index
when a negative skip is used and nothing is found.  Rather than
change the old routine in the same commit this defers the issue, given
that no working code in existence does the negative skip (by
definition) because of a desire to review if negative skip counts are
systemically a desirable thing to be concerned with when /SKIP is
offered as a refinement.