Unify PUSH_TRAP with stack/state balance check
There were several macros for detecting leaks, or imbalanced situations
in the stack.  These would be called at points where various things
were expected not to accumulate.

It became apparent over time that the variables being snapshotted and
compared against were the same ones that needed to be handled for leaks
by the setjmp/longjmp wrappers.  So this ties together the code for
pushing interpreter state to restore in case of an error with code
for snapshotting the interpreter state just to check that it has
balanced out.  This localizes the issues and also makes the debug
asserts more heavyweight, and gets rid of piecewise checks like
MANUALS_LEAK_CHECK_DEBUG.

Additionally this renames REBOL_STATE to just `struct Reb_State`, to
make it a bit less opaque what's going on.