Move POISON/UNPOISON macros to reb-c.h
The address sanitizer memory poisoning macros were private to the
%m-pools.c file, but can be useful to other parts of the code.  As
it is a "C level / platform" thing (kind of like the attributes) then
%reb-c.h seems as good a place as any to put them.

This also renames them to a more generic POISON_MEMORY/UNPOISON_MEMORY
as it is conceivable that there could eb other providers of the
functionality than Address Sanitizer.  (I've written a little C++
version that does XOR scrambling, actually.  Though not as good as
having a hook into memory accesses, it would work on compilers that
didn't have address sanitizer...for whatever that is worth.)

Also adds comments explaining what the poisoning is for, in case
someone reading hasn't heard of it.