New BACKTRACE native (replaces STACK)
The STACK routine that produced a backtrace was designed mainly for
reporting a block list of labels for an ERROR! to report its "where"
location.  The simple functionality and confusing interface wasn't
fitting for interactive stack exploration.

This implements a more featured version of BACKTRACE, which by default
is zero arity and gives numbered stack frames...suitable for an
interactive console that might be suspended in mid-stack (as in at a
debug break).  It attempts to also cover enough behaviors to implement
the previous STACK function's behavior, which is rewritten as
userspace code in `<r3-legacy>`.

For a version of the bare-bones functionality that just gives back
a list, BACKTRACE/ONLY provides that functionality at the moment.
Names and interface are very likely to change.