Get rid of unused Copy_Stack_Values
The Copy_Stack_Values() routine implements the same essential logic of
Pop_Stack_Values().  The difference is it's not used (and hence not
being exercised) and that it does not do the DS_DROP to remove the
items from the stack.

If this functionality were needed, it would make more sense to move
the DS_DROP out of Pop_Stack_Values and let those clients that wished
to drop do so, while those who merely wanted to copy would do that.
Then there would be one code path for the functionality.