Bump minimum Windows version to XP
The `#define _WIN32_WINNT 0x0500` in host-main.c specifies that the
version of the API being targeted is Windows 2000.  Microsoft Visual
Studio 2015 will give a warning/error on this, as its minimum supported
version is Windows XP (0x0501).

If building for Windows 2000 becomes a frequent issue, this could have
a compiler switch.  But for Ren-C's purposes it's better to not pick
something that the current version of the vendor's platform tools
does not support - and let those building for Windows 2000 roll it
back if they must.