Tweak host-main.c to use sys-core.h
This removes the dependency of %host-main.c (which implements the REPL)
on the RL_Api.  Instead it is able to use essentially any function that
the interpreter has established for its own use.  This gives it the
ability to inspect REBVALs or REBSERs, to handle THROW and FAIL, etc.

It's accomplished in a fairly minimalist change, which shifts the
responsibility of linking the host table off to %host-core.c - which
is pretty much entirely RL_Api and RXIARG calls (as currently written).

The change is integral to being able to adapt the console to support
interactive debugging (in an era prior to the definition and design
of an RL_Api replacement).