Win fixes, back out D_OUT => OUT change
This addresses some warnings in the Windows build related to macro
redefinitions that showed up when trying to include both <windows.h>
and the %sys-core.h, outside of the usual way sys-core includes
windows for its own purposes.

Apparently IN and OUT are defined in the Windows header files, so it
is likely not a good idea to use OUT as the macro name for the output
slot for natives to write to.  D_OUT is still the primary name and
only a few exploratory replacements were done, so this reverts them.

It also undefines IS_ERROR to avoid a compiler warning, and puts an
explanation of what the upshot is regarding including more windows
headers.

This fixes a warning about wide character lengths on windows as well.