Check function parameter read-only status
In Set_Var there was a check for whether or not the keys of an object
were locked, but this check was not done on the params of a function
frame.  Hence a protected function arg or local could be modified
despite that protection.

This adds the check to Set_Var.