Renamings to clarify _Deep, _Shallow
By convention, functions which operate "deeply" on arrays end in _Deep
to help convey their operation, while ones that only run on the top
level end in _Shallow.  The function Bind_Relative was deep but
didn't say so (some comment next to one said "//!! needs deep", which
suggests some kind of confusion).  So this adds that.

Also in the ordering, _Shallow or _Deep is the last name part (other
than a potential _Managed).  The Copy_Array_At_Shallow_Extra had the
"Extra" (which corresponds to an 'argument') before the _Shallow.
Reordered for consistency with the other routines in the same scheme.