Remove superfluous ASSERT_BALANCED
The main loop in Do_Core uses SNAP_STATE and a later ASSERT_BALANCED
on each loop iteration to make sure nothing accrues (manual series
created but not freed or transitioned to managed, etc.)  It is
checked at the end of every evaluative step.  But it was also checked
at the start of every step.  This is unnecessarily redundant as the
only state it would verify that checking at the end doesn't already
would be that ASSERT_BALANCED immediately after a SNAP_STATE is
balanced...which would almost certainly raise an alarm in other ways.

Though this only affects the debug build, there's no reason to make
it check twice for no reason on every cycle.

This also moves the tail check ASSERT_BALANCED so that it does its
check before the length check (immediately after the switch) so that
the check is not skipped when the index goes past the length of the
array being processed.