Update expr_index marker if reusing lookahead work
A Ren-C optimization allows a non-/NEXT-oriented Do to reuse the word
lookup done by infix lookahead as a "pre-fetch" for that word to use
in the next evaluation step.

However, this update was not refreshing the "begin of expression"
index (a new Ren-C concept for giving better debug information).  This
updates the expr_index.

There are other initializations that help the debugger which should
probably be factored into a shared routine at some point, so that
these lookaheads get the same debugging coverage as ordinary indexing
does.  Until that time, a quick compromise uses the SPORADICALLY()
macro to run both paths...the shortcut and the non-shortcut, split
half the time each way.