Read-only binding to NATIVE!/ACTION! params
This is a speculative--and possibly bad--idea, which is to permit user
code to bind to the arguments of natives.

The motivation is to allow debugging routines to introspect the state
of the routines, which may include deliberate exposure of state as
local variables.  (PARSE, for instance, might keep its parse position
in a POSITION: variable in the frame rather than hidden internally,
thus allowing someone viewing the stack to see where it currently is.)

For efficiency, natives will often modify their arguments, as this
allows them to be protected from garbage collection.  This practice
may be less necessary, as the new design for DO state would allow
that protection to come from pointing the ARG, PARAM, and REFINE
pointers at any local memory cell...and the D_CELL protected memory is
offered.

Whether this feature stays or not is open for debate, but for the
beginnings of debugging it is assumed to be desirable to be able to
bind to the native args for purposes of introspection.  Those not
using the debugger should not be affected.