Factor 2 debug-only functions out of Do_Core
Do_Core() is a very long function.  This isn't a problem in and of
itself (as it contains no repeated portions or code that would be
useful elsewhere).  However there actually was a bit of debug code
that needed to be factored out, as it should have been called from
two different places.

Given that the overhead of a function call in the middle of the
evaluator isn't an issue if that function is debug-only, this breaks
out two debug prep routines--one that runs only once at the beginning
of Do_Core, and one that runs on each evaluation step in order to
ensure there's no contamination between runs.  This also resolved an
issue cited in an earlier commit about the infix fetch that reuses
work via `do_fetched_word` not doing the debug work, and also that it
was not incrementing the debug counter for the do_index to use
in breakpoints.

This also adds some refactoring and comments, which caught statements
that contradicted about the responsibility of SET_TRASH_SAFE on the
output slot on each iteration.  The behvior that is currently working
is preserved but notes are left to review it further to see if it is
setting it more than it needs to.