Programmatic GDB breakpoint support
Not to be confused with the current work on breakpoints in the Rebol
code itself, this allows a parallel facility to add programmatic
breakpoints into the C code for Rebol.  It does so by adding in a
header called %debugbreak.h from this GitHub repo:

https://github.com/scottt/debugbreak

It includes an experimental usage in the "break on tick count" code
(which should be triggered by a command line option and not require
a recompilation).

The file says it has a fallback path for unsupported architectures,
so this is to run through the build configurations and see if that
is indeed the case.