EXIT/FROM any stack level, natives
This alters the behavior of dynamically-scoped EXIT in two significant
ways.  One is that natives will be exited just like non-natives, so
if you were to write:

    foo: func [] [if 1 < 2 [exit print "no"] print "yes"]

Then calling foo would print "yes", because IF would respond to the
EXIT signal.

This is coupled with another change which is to allow the targeting of
a specific function or closure to exit via the /FROM refinement.  A
name of a function or the frame object of a closure can be used.  So
this exit will reach past the IF to WHILE to break the infinite loop.

    while [true] [if 1 < 2 [exit/from :while]]

Stronger contracts will be possible with FUNCTION! (as with CLOSURE!)
when each function frame is individually distinguishable.  In the
meantime, the most recent call on the stack will receive the exit
command.