Make WHERE primary backtrace info, factor code
To make the backtrace more user friendly, this switches it to show
more information by default than just the label of the call.
It gives a snapshot of the expression at the WHERE spot.  Also,
pending stack frames are included to show the impetus for an expression
evaluation, even if that isn't being run yet.

This also breaks out two service functions from the backtrace, to turn
indices the backtrace presents back into call frames...and to generate
the subsetted WHERE block for a stacklevel.