Breakpoints + Interactive Debugging
This is a start on interactive debugging.  The means of entry to the
interactive debugging mode are BREAKPOINT the native, and a now Ctrl-C
will enter the debugging mode as well.

(Note: There are several usability questions and questions of what it
means to "play nice" as a console program.  Many scripts might desire
Ctrl-C to exit and not enter the debugger.  Both desires are valid,
and the decisions about configuration are a separate concern from the
groundwork to be able to make the choice.  For now, Ctrl-C is chosen
to break; and QUIT + HALT are both options for reacting to it.)

The commands RESUME, BACKTRACE, and DEBUG are used to work with the
interrupted state.  These commands are evolving and will be documented
on the Trello.