Remove debug_break() test
The test seems to have been successful in terms of the build variants.
While having the code automatically do a debug break would be useful
if requested at the command line, it's a bit of a liability as if
one is not running under a debugger then it halts the program instead
of getting more debug output.  So remove for the moment.