Convert canon NONE, UNSET, etc. to globals
The pre-cooked values for NONE and UNSET were living in the ROOT
object, despite not needing any garbage collector awareness.  This
isn't as good as having them be out of the root object, because then
it would be possible to trap array accesses to them by following
them with a trash value in the debug build...as opposed to stray
clients receiving the isolated value doing a ++ and finding some other
value in the root object.

This does that protection, as well as add a canon FALSE and TRUE.