RESET_HEADER before zeroing in SET_ZEROED()
The questionable concept of SET_ZEROED--a method used by ZERO? and
POSITIVE? and NEGATIVE?--of creating a zero value was to clear the
bits and set the header, assuming zero bits equals value zero.
(Pulling the R3-Alpha behavior out into a routine and naming was
done in Ren-C.)

The previous implementation would zero a full REBVAL size slot, and
then reset the header.  Micro-optimization-wise this ends up zeroing
the header and then overwriting it, but more importantly it doesn't
allow RESET_HEADER to inspect what bits were there before to (for
instance) know if this is a legal value slot to overwrite.

This adjusts it to reset the header first, and then only zeros the
bits in the payload...leaving the header as-is.