Retake `out` to mean `function out param`
The emerging C convention of routines that return a value by writing it
into a pointer passed as an argument (as opposed as a result via
`return`) is to use `out` either as the name for the argument, or
a suffix `_out`.  This is a light cleanup on that issue so that
local variables which are not being passed back to the caller are
not labeled with out.

`result` is used, because what these meant in general was to store
the result of a call to DO...and this is the convention that had
been established for that elsewhere.  However, that is contentious
with another possible use of `result` to be a temporary local variable
for holding what will ultimately be the return value.  `ret` is
another possibility for that, but these try to use names that identify
what the returned value represents instead.