Make empty payload tracking an #ifdef
The 4 types with empty payloads (TRASH, UNSET, NONE, and LOGIC) have
the opportunity to record where they originated in the debug build,
as well as the tick count.  This is a very useful debugging feature,
but it does come with a drawback--it initializes memory addresses
with data that might otherwise be uninitialized, so accesses could
be more readily caught by tracking tools.

This explains the tradeoff and adds a #define for it (though currently
defaults it to being defined.