Pre-format all REBVAL cells for writability
This commit turns on the feature of paying attention to the WRITABLE
bit in a REBVAL header.  This makes it possible to enable and disable
writes to arbitrary REBVAL cells in the debug build.

In order for this to work, REBVALs living in series or on the stack
must have the writable bit set if code intends to set those values.
It is therefore no longer possible to write just:

    REBVAL value;
    SET_INTEGER(&value, 1020);

Allowing SET_INTEGER to overwrite arbitrary memory would undermine
the purpose of the "writable" bit.  Hence it must be safely cleared
in debug builds:

    REBVAL value;
    VAL_INIT_WRITABLE_DEBUG(&value);
    SET_INTEGER(&value, 1020);

While it is slightly irritating to have to do this, most values live
in series so it's taken care of.  For the purposes of a system-wide
audit of stack-based REBVALs, the init writable was rearranged to
be as close to the declaration as possible (usually right under it).

This is a first take on the implementation, so it doesn't do a
fine-grained formatting of series where the extra space (capacity and
bias) are considered read-only and require an expand before they are
written.  That will involve more work, but this does at least check
the bounds of the capacity of the series.