Make BACKTRACE print the output by default
Returning the result as a value means that in the REPL it gets cut
off, and the backtrace was designed to have the most relevant stack
frames appear at the end (to be closer to the REPL input).  This
changes it to print by default and makes /ONLY return the value.

The previous /ONLY refinement is renamed as /BRIEF.