Make smoke test execution run once with output
The simple test to spin-up the interpreter and see if it can print
"OK" pipes the output through grep.  If it fails, all you see is
"didn't print OK".  It can be useful to know what it said or did
instead, especially with situations that happen only on Travis but
cannot be recreated locally.

This does a second spin-up that doesn't go through grep before the
one that does, to hopefully log some of that output.