Pending frames count as "first" in backtrace
The Call_For_Stack_Level_Function() repeats logic in backtrace for
routines that wish to follow up on and use the numbers that backtrace
gave out.  As a sanity check, backtrace makes sure the call that
would be given back for a number it hands out line up.

This was failing because pending frames were not being counted as
"first" occurrences for the test against breakpoint.