Fix FFI callback value initialization
A transition made it so that a function's paramlist (the FUNC) was able
to completely convey the entire function's REBVAL, by virtue of storing
a copy of the canon REBVAL in the [0] slot of that parameter list.

(This is a technique very similar to what R3-Alpha called the internal
type "FRAME!" in objects, a parallel one to which has replaced the
R3-Alpha "FRAME!" by putting ANY-CONTEXT! values like OBJECT! as
canon values in the zero slot of the varlists of those contexts.)

A bug in the transitioning for FFI would only initialize the paramlist
field of a callback, skipping the body and spec initialization from
the canon value.  This retrieves the full canon value, avoiding leaving
the spec and body uninitialized when extracted.