FRAME => CONTEXT split (most uses to CONTEXT)
The word "frame" classically evokes "call frame" for a function.  In
R3-Alpha, the word was used to refer to the keys and values of an
object...usually.  It had a strange associated type called FRAME! that
was not intended to be exposed to users, which allowed a single
series to store some "extra" information in the first value slot, so
that only the series pointer need be passed around...yet have more
"meta-properties" of that series to extract.

That usage of FRAME! has been replaced in Ren-C, and the term is
about to be re-purposed as the programmatic exposure of the call frame
of a specific function instance.  Currently that is only technically
possible to do in a GC-safe way with a CLOSURE!...but the goal is
to push the limits of FUNCTION! until it can do the same (in a way
that is "affordable", compared to CLOSURE!'s relatively high costs
per-instantiation).

This FRAME! type will for the most part behave to the user's perception
like an OBJECT!.  Main differences will be that it will not be
expandable (as OBJECT! may not be by default, either).  But also that
when the function ends, any words bound to the frame will expire
permanently.  Nevertheless, it will be considered to be an ANY-CONTEXT!
which is the type category in which ERROR!, OBJECT!, PORT!, MODULE!
are members.

In the renaming here, most things went in the direction of being
called CONTEXT, but anything that is going to target this new FRAME
was kept as frame.  It's not necessarily the end point, and some of
the macros + variable names are pretty verbose, as in:

    CONTEXT_VARS_HEAD(CONTEXT_VALUE(VAL_CONTEXT(...))...

Some abbreviation may be in order, though perhaps better to start
abbreviating variables (c instead of context) rather than obscure the
macros.  This will all be reviewed later.