Delete Set_Var (redundant with GET_MUTABLE_VAR)
The Get_Var primitive was enhanced in Ren-C to provide both mutable
and const forms, sharing the same code path.  (Macros do the work to
differentiate the interface.)  It returns a pointer to the variable
slot and manages the protections consistent with the const status.

Set_Var is thus redundant and a subset.  Because Set_Var takes a
REBVAL and bit copies it, then it is doing nothing that you could
not just do with:

    *GET_MUTABLE_VAR(any_word) = *value;

However, GET_MUTABLE_VAR() offers back a pointer, which can be set
more efficienly via macros than by full assignment.  For instance:

    SET_NONE(GET_MUTABLE_VAR(any_word));

The SET_NONE macro only sets the header, so it does 1/4 the writing
that setting to an arbitrary REBVAL that happened-to-be-none would do,
because no optimization would be made (it would assume the worst case
scenario that all the payload bits are significant).